link: move btf_id.go into tracing.go
